refactor: useSetReminder에  mutation 로직 분리 및 ReminderModalButtons component 제거 후 ReminderModal component로 합성
