+++ frontend/src/hooks/useSetReminder.ts
@@ -1,60 +1,7 @@
-import { getDateInformation, getMeridiemTime, ISOConverter } from "@src/@utils";
-import { deleteReminder, postReminder, putReminder } from "@src/api/reminders";
+import { getDateInformation, getMeridiemTime } from "@src/@utils";
 import { useEffect, useRef } from "react";
-import { useMutation } from "react-query";
 import useDropdown from "./useDropdown";
 import useInput from "./useInput";
-import useSnackbar from "./useSnackbar";
-
-interface IsInvalidateDateTimeProps {
-  checkedYear: number;
-  checkedMonth: number;
-  checkedDate: number;
-  checkedHour: number;
-  checkedMinute: number;
-  year: number;
-  month: number;
-  date: number;
-  hour: number;
-  minute: number;
-}
-
-const isInvalidateDateTime = ({
-  checkedYear,
-  checkedMonth,
-  checkedDate,
-  checkedHour,
-  checkedMinute,
-  year,
-  month,
-  date,
-  hour,
-  minute,
-}: IsInvalidateDateTimeProps) => {
-  if (checkedYear < year) return true;
-  if (checkedYear <= year && checkedMonth < month) return true;
-  if (checkedYear <= year && checkedMonth <= month && checkedDate < date)
-    return true;
-
-  if (
-    checkedYear <= year &&
-    checkedMonth <= month &&
-    checkedDate <= date &&
-    checkedHour < hour
-  )
-    return true;
-
-  if (
-    checkedYear <= year &&
-    checkedMonth <= month &&
-    checkedDate <= date &&
-    checkedHour <= hour &&
-    checkedMinute < minute
-  )
-    return true;
-
-  return false;
-};
 
 const convertTimeToStepTenMinuteTime = ({
   hour,
@@ -68,17 +15,6 @@ const convertTimeToStepTenMinuteTime = ({
   }
 
   return { parsedHour: hour, parsedMinute: Math.ceil(minute / 10) * 10 };
-};
-
-const convertMeridiemHourToStandardHour = (
-  meridiem: string,
-  meridiemHour: number
-): number => {
-  if (meridiem === "오후") {
-    return meridiemHour === 12 ? 0 : meridiemHour + 12;
-  }
-
-  return meridiemHour;
 };
 
 const parsedDateTime = (ISODateTime: string) => {
@@ -104,32 +40,10 @@ const invalidMeridiem = (value: string) => {
 };
 
 interface Props {
-  messageId: string;
   remindDate: string;
-  handleCloseReminderModal: () => void;
-  refetchFeed: () => void;
 }
 
-function useSetReminder({
-  messageId,
-  remindDate,
-  handleCloseReminderModal,
-  refetchFeed,
-}: Props) {
-  const { mutate: addReminder } = useMutation(postReminder, {
-    onSuccess: () => {
-      handleCloseReminderModal();
-      refetchFeed();
-    },
-  });
-
-  const { mutate: modifyReminder } = useMutation(putReminder, {
-    onSuccess: () => {
-      handleCloseReminderModal();
-      refetchFeed();
-    },
-  });
-
+function useSetReminder({ remindDate }: Props) {
   const { year, month, date, hour, minute } = getDateInformation(new Date());
   const { date: lastDate } = getDateInformation(new Date(year, month, 0));
   const { meridiem, hour: meridiemHour } = getMeridiemTime(hour);
@@ -150,7 +64,6 @@ function useSetReminder({
     isDropdownOpened: isDateTimePickerOpened,
     handleToggleDropdown: handleToggleDateTimePicker,
   } = useDropdown();
-  const { openFailureSnackbar } = useSnackbar();
 
   const {
     value: checkedYear,
@@ -209,90 +122,11 @@ function useSetReminder({
   const replaceCheckedMonth = Number(checkedMonth.replace("월", ""));
   const replaceCheckedDate = Number(checkedDate.replace("일", ""));
 
-  const replaceCheckedHour = convertMeridiemHourToStandardHour(
-    checkedMeridiem,
-    Number(checkedHour.replace("시", ""))
-  );
+  // const replaceCheckedHour = convertMeridiemHourToStandardHour(
+  //   checkedMeridiem,
+  //   Number(checkedHour.replace("시", ""))
+  // );
   const replaceCheckedMinute = Number(checkedMinute.replace("분", ""));
-
-  const handleCreateReminder = () => {
-    if (
-      isInvalidateDateTime({
-        checkedYear: replaceCheckedYear,
-        checkedMonth: replaceCheckedMonth,
-        checkedDate: replaceCheckedDate,
-        checkedHour: replaceCheckedHour,
-        checkedMinute: replaceCheckedMinute,
-        year,
-        month,
-        date,
-        hour,
-        minute,
-      })
-    ) {
-      openFailureSnackbar(
-        "리마인더 시간은 현재 시간보다 미래로 설정해주셔야 합니다."
-      );
-
-      return;
-    }
-
-    const reminderISODateTime = ISOConverter(
-      `${replaceCheckedYear}-${replaceCheckedMonth}-${replaceCheckedDate}`,
-      `${replaceCheckedHour}:${replaceCheckedMinute}`
-    );
-
-    addReminder({
-      messageId: Number(messageId),
-      reminderDate: reminderISODateTime,
-    });
-  };
-
-  const handleModifyReminder = () => {
-    if (
-      isInvalidateDateTime({
-        checkedYear: replaceCheckedYear,
-        checkedMonth: replaceCheckedMonth,
-        checkedDate: replaceCheckedDate,
-        checkedHour: replaceCheckedHour,
-        checkedMinute: replaceCheckedMinute,
-        year,
-        month,
-        date,
-        hour,
-        minute,
-      })
-    ) {
-      openFailureSnackbar(
-        "리마인더 시간은 현재 시간보다 미래로 설정해주셔야 합니다."
-      );
-
-      return;
-    }
-
-    const reminderISODateTime = ISOConverter(
-      `${replaceCheckedYear}-${replaceCheckedMonth}-${replaceCheckedDate}`,
-      `${replaceCheckedHour}:${replaceCheckedMinute}`
-    );
-
-    modifyReminder({
-      messageId: Number(messageId),
-      reminderDate: reminderISODateTime,
-    });
-  };
-
-  const handleRemoveReminder = async (messageId: string) => {
-    if (window.confirm("해당하는 메시지 리마인더를 정말 삭제하시겠습니까?")) {
-      await deleteReminder(messageId);
-
-      refetchFeed();
-      handleCloseReminderModal();
-
-      return;
-    }
-
-    return;
-  };
 
   useEffect(() => {
     if (remindDate) {
@@ -394,9 +228,6 @@ function useSetReminder({
       handleChangeMonth,
       handleChangeDate,
       handleToggleDateTimePicker,
-      handleCreateReminder,
-      handleModifyReminder,
-      handleRemoveReminder,
     },
   };
 }
